ALGO : small fixes nth_element impl

+++ Divide_and_conquer/nth_element.h
@@ -19,7 +19,7 @@ namespace divide_and_conquer_ns
 	    auto it_greater = std::partition(it_beg, it_end, [&pivot](const auto& value) { return value <= pivot; });
 	    auto it_equal = std::partition(it_beg, it_greater, [&pivot](const auto& value) { return value < pivot; });
 	    if (it_beg + i >= it_greater)
-		return nth_element_impl(it_greater, it_end, it_beg + i - it_greater);
+		return nth_element_impl(it_greater, it_end, static_cast<int>(it_beg + i - it_greater));
 	    else if (it_beg + i < it_equal)
 		return nth_element_impl(it_beg, it_equal, i);
 	    else
